Add files via upload
In this lab, I learned how to use the Filter and Sort tools in Excel to filter and sort our data to enable us to control what information is displayed, and how it is displayed in our worksheets. Next, I learned how to use some of the most common functions a Data Analyst might use; namely IF, IFS, COUNTIF, and SUMIF. Finally, I learned how to use the VLOOKUP and HLOOKUP functions in Excel to reference data contained in both vertical and horizontal lookup tables.
All of this is demonstrated in these files.